requirements.txt
SpeechRecognition==3.8.1
numpy==1.16.0
requests==2.11.1
InstagramAPI==1.0.2
Pillow==6.1.0
